+++ test/container/container_test.dart
@@ -33,4 +33,25 @@ void main() {
       expect(container.make('key'), null);
     });
   });
+
+  group('clone', () {
+    test('Clones should have access to the bindings declare before cloning',
+        () {
+      var container = CacheContainer();
+      container.bind('foo', (p0) => 'bar');
+
+      var container2 = container.clone();
+      expect(container2.make('foo'), equals('bar'));
+    });
+
+    test(
+        'updating a binding in a clone should not update the binding for the root',
+        () {
+      var container = CacheContainer();
+      var container2 = container.clone();
+
+      container2.bind('foo', (p0) => 'bar');
+      expect(container.make('foo'), equals(null));
+    });
+  });
 }